Make the fetch-data script show a diff of the runs
I created a little Levenshtein distance function for calculating
the best diff between the new and old list of runs. This will
make it less unknowable what is changing when I update the runs
list.

I also added a prompt of whether I'd like to overwrite the file
(not sure why, just in case something horrible happens maybe) and
also an automated backup of the runs that I download so that I
don't have to remember to manually make copies of the old schedules
when there's a new event.